+++ index.html
@@ -24,6 +24,6 @@
 			>
 		</noscript>
 		<div id="app"></div>
-		<script type="module" src="src/options/options.ts"></script>
+		<script type="module" src="src/app/options/options.ts"></script>
 	</body>
 </html>
