+++ src/app/chat/Chat.vue
@@ -1,0 +1,7 @@
+<template>
+	<div />
+</template>
+
+<script setup lang="ts">
+void 0;
+</script>
