+++ src/app/options/views/Onboarding/Onboarding.vue
