+++ src/app/options/views/Onboarding/OnboardingCompat.vue
