+++ src/app/options/views/Onboarding/OnboardingConfig.vue
