+++ src/app/options/views/Onboarding/OnboardingEnd.vue
