+++ src/app/options/views/Onboarding/OnboardingPlatforms.vue
