+++ src/app/options/views/Onboarding/OnboardingPromo.vue
