+++ src/app/options/views/Onboarding/OnboardingStart.vue
